+++ mlaut/benchmarking/data/__init__.py
@@ -1,0 +1,4 @@
+from ._base import BaseDataset
+from ._ram import RAMDataset
+from ._hdd import HDDBaseDataset
+
