+++ mlaut/benchmarking/results/__init__.py
@@ -1,0 +1,3 @@
+from mlaut.benchmarking.results._base import BaseResults, HDDBaseResults, _PredictionsWrapper
+from mlaut.benchmarking.results._hdd import HDDResults
+from mlaut.benchmarking.results._ram import RAMResults
