+++ setup.py
@@ -11,6 +11,8 @@ setup(name='mlaut',
       license='BDS',
       packages=['mlaut',
                 'mlaut.benchmarking',
+                'mlaut.benchmarking.data',
+                'mlaut.benchmarking.results',
                 'mlaut.experiments',
                 'mlaut.highlevel',
                 'mlaut.resampling',
